Section component moved to common directory
